Justerer komponenter for innlogget mobilmeny
