+++ src/komponenter/header/header-regular/common/minside-lock-msg/MinsideLockMsg.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../../styling-variabler.less';
+@import (reference) 'src/styling-variabler.less';
 
 .minside-lock {
     &__panel {
@@ -12,7 +12,7 @@
         border: 2px solid @navds-global-color-lightblue-500;
 
         @media @screen-mobile {
-            margin: 0;
+            margin-bottom: 1rem;
         }
     }
