+++ src/komponenter/header/header-regular/mobil/meny/innhold/hovedmeny/innlogget/DittNavMeny.less
@@ -1,0 +1,4 @@
+.mobilDittNavMeny {
+    border-bottom: 1px solid var(--navds-global-color-gray-600);
+    padding-bottom: 1rem;
+}
