+++ src/komponenter/header/header-regular/mobil/meny/innhold/hovedmeny/innlogget/MobilInnloggetBruker.less
@@ -2,28 +2,14 @@
 
 .innloggetbruker {
     padding: 1rem 0 2.5rem 0;
+
     &__bruker {
-        color: @navds-global-color-blue-500;
         text-transform: capitalize;
     }
 
     &__loggut {
-        padding-top: 0.5rem;
-        display: flex;
-        align-items: center;
-        background: none;
-        border: none;
-
-        &:focus {
-            outline: 2px solid @navds-global-color-blue-500;
-        }
-
-        &--text {
-            padding: 0 0 0 0.75rem;
-
-            .navds-ingress {
-                color: @navds-global-color-blue-500;
-            }
-        }
+        margin-top: 0.5rem;
+        margin-left: -0.5rem;
+        font-weight: var(--navds-font-weight-bold);
     }
 }
