+++ src/komponenter/header/header-regular/mobil/meny/innhold/hovedmeny/innlogget/MobilInnloggetForsideLenke.less
@@ -1,0 +1,5 @@
+.mobilInnloggetForsideLenke {
+    padding: 0 1rem 1rem 1rem;
+    margin-left: -1rem;
+    margin-right: auto;
+}
