+++ src/komponenter/header/header-regular/mobil/meny/innhold/hovedmeny/menypunkt/MobilMenypunkt.less
@@ -1,6 +1,5 @@
 .mobilMenyLenke {
     display: flex;
-    justify-content: space-between;
     align-items: center;
     text-decoration: none;
     padding: 1rem;
@@ -8,6 +7,10 @@
 
     font-size: var(--navds-font-size-heading-small);
     font-weight: var(--navds-font-weight-bold);
+
+    & > svg {
+        margin-left: auto;
+    }
 
     &:hover {
         text-decoration: underline;
