+++ src/komponenter/header/header-regular/mobil/meny/innhold/hovedmeny/menypunkt/MobilMenypunkt.tsx
